Adapt imports after gammapy.irf.psf_gauss rename
