+++ gammapy/irf/__init__.py
@@ -8,7 +8,7 @@ from .psf_core import *
 from .psf_table import *
 from .psf_3d import *
 from .psf_check import *
-from .psf_analytical import *
+from .psf_gauss import *
 from .psf_king import *
 from .energy_dispersion import *
 from .irf_stack import *
